Rename namespce Cnab.Bundle to Cnab to avoid symbol name duplication
